Added a typedef for a basic namespace descriptor, which downcasts from generator...

+++ include/event_model/home/types/2_int8.h
@@ -14,7 +14,7 @@ struct rbl_type_type_traits<rbl_int8>
   {
     const char * operator()()
     {
-      return "int";
+      return "bigint";
     }
   };
